CHE-8164: Fix error when creating project from archetype (#10029)
When creating a project with the help of Maven Archetype Plugin it names project directory by artifact-id. If artifact-id != project name specified in the first page of the project wizard, error happens. Maven Archetype Plugin has basedir property that can solve this problem, but there is a bug in the plugin: https://issues.apache.org/jira/browse/ARCHETYPE-311. Rename the project directory to specified project name solves the problem.